* fixed *nix x64 support, define X64 in config.h until I figure out how to do it from autoconf.
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2166 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/ScriptMgr.cpp
@@ -113,7 +113,7 @@ char *ext;
 			if (ext != NULL && !strcmp(ext, ".so")) {
 				string full_path = "../lib/" + string(list[filecount]->d_name);
 				SCRIPT_MODULE mod = dlopen(full_path.c_str(), RTLD_NOW);
-				printf("  %s : 0x%08X : ", list[filecount]->d_name, (unsigned int)mod);
+				printf("  %s : 0x%08X : ", list[filecount]->d_name, (unsigned long)mod);
 				if(mod == 0)
 					printf("error! [%s]\n", dlerror());
 				else
